move Scrolling icon and function into seperate component
