+++ src/style.css
@@ -24,7 +24,7 @@ canvas {
   align-items: center;
   left: 0;
   right: 0;
-  bottom: 20px;
+  bottom: 25px;
 }
 .scroll-container > * {
   color: rgba(255, 255, 255, 0.654);
